add validation and test_user_actions

+++ app/views.py
@@ -14,7 +14,8 @@ section = 'postgresql'
 user_actions_object = UserActions(path, section)
 
 app = create_app()
-app.config['SECRET_KEY'] = 'secret123'
+app.config['SECRET_KEY'] = 'thismustbesecrete'
+
 
 def token_required(f):
     @wraps(f)
@@ -22,10 +23,10 @@ def token_required(f):
         token = None
         if 'x-access-token' in request.headers:
             token = request.headers['x-access-token']
-        
+
         if not token:
             return jsonify({'message': 'Token is missing'}), 401
-        
+
         try:
             data = jwt.decode(token, app.config['SECRET_KEY'])
             current_use = user_actions_object.get_user_by_id(data['user_id'])
@@ -35,35 +36,41 @@ def token_required(f):
         return f(current_user, *args, **kwargs)
     return decorated
 
+
 @app.errorhandler(404)
 def not_found(error):
     return make_response(jsonify({'error': 'Resource not found'}), 404)
 
+
 def validate_answer_object(request_object):
     if not request_object:
-        return jsonify({'error': 'improper data format',
-                        'help - proper format': {
-                            'answer_body': 'answer body'
-                        }}), 400
+        return jsonify({'error': 'unsuported format'}), 400
     if 'answer_body' not in request_object:
-        return {'error': 'please answer can not be empty'}
+        return {'error': 'please answer can not be empty'}, 400
 
 # user signup endpoint
+
+
 @app.route('/api/v2/auth/signup', methods=['POST'])
 def register_user():
     request_data = request.get_json()
     user_email = request_data['user_email']
     user_password = request_data['user_password']
-    reg = user_actions_object.user_register(user_email, user_password)
-    return jsonify(reg), 201
+    if not validate_email(user_email):
+        user_actions_object.user_register(user_email, user_password)
+        return jsonify({'success': 'registered successfully'}), 201
+    else:
+        return validate_email(user_email)
 
 # user signin endpoint
+
+
 @app.route('/api/v2/auth/login', methods=['GET'])
 def login_user():
     auth = request.authorization
     if not auth or not auth.username or not auth.password:
-        return make_response('unauthorized accessss', 401, {'WWW-Authenticate': 
-        'Basic realm="Login required!"'})
+        return make_response('unauthorized accessss', 401, {'WWW-Authenticate':
+                                                            'Basic realm="Login required!"'})
     user = user_actions_object.user_login(auth.username)
     if not user:
         return make_response('Could not verify', 401, {'WWW-Authenticate':
@@ -71,25 +78,33 @@ def login_user():
     if check_password_hash(user[3], auth.password):
         token = jwt.encode(
             {
-                'user_id': user[0], 'exp' :datetime.datetime.utcnow() + datetime.timedelta(
+                'user_id': user[0], 'exp': datetime.datetime.utcnow() + datetime.timedelta(
                     minutes=20
                 )
             }, app.config['SECRET_KEY']
         )
-        return jsonify({'token' : token.decode('UTF-8')})
+        return jsonify({'token': token.decode('UTF-8')})
     return make_response('unauthorized access', 401, {'WWW-Authenticate':
                                                       'Basic realm="Login required!"'})
-    
+
 # create question endpoint
+
+
 @app.route('/api/v2/question', methods=['POST'])
 @token_required
 def post_question(current_user):
     request_data = request.get_json()
-    if not validate_question_object(request_data):
-        title = request_data['question_title']
-        body = request_data['question_body']
+    title = request_data['question_title']
+    body = request_data['question_body']
+    if not validate_question_object(request_data) and not validate_question(title):
         user_actions_object.create_question(current_user, title, body)
-    return jsonify({'message': 'question successfully created'}), 201
+        return jsonify({'message': 'question successfully created'}), 201
+    else:
+        check = {
+            'status': validate_question_object(request_data),
+            'status': validate_question(title)
+        }
+        return jsonify({'check_status': check})
 
 def validate_question_object(request_object):
     if not request_object:
@@ -105,13 +120,15 @@ def validate_question_object(request_object):
         return jsonify({'error': 'please body is required'}), 400
 
 # get all questions endpoint
+
+
 @app.route('/api/v2/questions', methods=['GET'])
 def get_all():
     results = user_actions_object.view_all_questions()
     container = []
     for result in results:
         q_obj = {
-            'question_id':result[0],
+            'question_id': result[0],
             'question_author_id': result[1],
             'question_title': result[2],
             'question_body': result[3]
@@ -120,6 +137,8 @@ def get_all():
     return jsonify({'All Questions': container}), 200
 
 # get single questions endpoint
+
+
 @app.route('/api/v2/question/<int:question_id>', methods=['GET'])
 def get_single_question(question_id):
     results = user_actions_object.view_single_question(question_id)
@@ -128,7 +147,7 @@ def get_single_question(question_id):
     answer_container = []
     for result in answers:
         a_obj = {
-            'answer_id':result[0],
+            'answer_id': result[0],
             'question_author_id': result[1],
             'question_id': result[2],
             'answer_body': result[3],
@@ -146,22 +165,44 @@ def get_single_question(question_id):
     return jsonify(container), 200
 
 # update question endpoint
+
+
 @app.route('/api/v2/question/<int:question_id>', methods=['PUT'])
-def update_question(question_id):
+@token_required
+def update_question(current_user, question_id):
+    get_one_question = user_actions_object.view_single_question(question_id)
     request_data = request.get_json()
-    if not validate_question_object(request_data):
-        title = request_data['question_title']
-        body = request_data['question_body']
-        user_actions_object.update_question(title, body, question_id)
-    return jsonify({'message': 'Question successfully updated'}), 200
+    title = request_data['question_title']
+    body = request_data['question_body']
+    if get_one_question[1] == current_user:
+        if not validate_question_object(request_data) and not validate_question(title):
+            user_actions_object.update_question(title, body, question_id)
+            return jsonify({'message': 'Question successfully updated'}), 200
+        else:
+            check = {
+                'status': validate_question_object(request_data),
+                'status': validate_question(title)
+            }
+            return jsonify({'check_status': check})
+    else:
+        return jsonify({'message': 'You are not the owner of the question'}), 401
 
 # delete question endpoint
+
+
 @app.route('/api/v2/question/<int:question_id>', methods=['DELETE'])
-def delete_question(question_id):
-    user_actions_object.delete_question(question_id)
-    return jsonify({'message': 'Question successfully deleted'}), 200
-    
+@token_required
+def delete_question(current_user, question_id):
+    get_one_question = user_actions_object.view_single_question(question_id)
+    if get_one_question[1] == current_user:
+        user_actions_object.delete_question(question_id)
+        return jsonify({'message': 'Question successfully deleted'}), 200
+    return jsonify({'error': 'You are not the owner of the question'}), 401
+
+
 # add answer endpoint
+
+
 @app.route('/api/v2/question/<int:question_id>/answer', methods=['POST'])
 @token_required
 def post_answer(current_user, question_id):
@@ -176,6 +217,8 @@ def post_answer(current_user, question_id):
         return jsonify({'message': 'Answer not added'}), 400
 
 # add update answer endpoint
+
+
 @app.route('/api/v2/question/<int:question_id>/answer/<int:answer_id>', methods=['PUT'])
 @token_required
 def upadte_answer(current_user, question_id, answer_id):
@@ -184,12 +227,13 @@ def upadte_answer(current_user, question_id, answer_id):
     if get_one_answer[1] == current_user:
         request_data = request.get_json()
         answer_body = request_data['answer_body']
-        update_answer = user_actions_object.update_answer(answer_id, answer_body)
+        update_answer = user_actions_object.update_answer(
+            answer_id, answer_body)
         if update_answer:
             return jsonify({'message': 'Answer successfully update'}), 201
         else:
             return jsonify({'message': 'Answer not Updated'}), 400
-    
+
     elif get_one_question[1] == current_user:
         update_answer_user = user_actions_object.update_answer_user(answer_id)
         if update_answer_user:
@@ -198,3 +242,47 @@ def upadte_answer(current_user, question_id, answer_id):
             return jsonify({'message': 'Answer not Updated'}), 400
     else:
         return jsonify({'message': 'You are not the author of the question or the answer'}), 401
+
+
+def validate_email(user_email):
+    check_duplicate_email = user_actions_object.get_user(user_email)
+
+    if check_duplicate_email > 0:
+        return jsonify({'message': 'Sorry user already exixts'}), 409
+    elif len(user_email) < 6:
+        return jsonify({'error': 'Email can not be less than six\
+            characters'}), 400
+    elif user_email.isdigit():
+        return jsonify({'error': 'Email format not allowed\
+        an email can not only have numbers'}), 400
+    elif "@" not in user_email:
+        return jsonify({'error': 'Email format not allowed\
+        an email must conatain @'}), 400
+    elif "." not in user_email:
+        return jsonify({'error': 'Email format not allowed\
+        an email must conatain . character'}), 400
+    elif user_email.startswith("@") or user_email.startswith("."):
+        return jsonify({'error': 'Email format not allowed\
+        an email must not start with @ or . character'}), 400
+    elif "@." in user_email:
+        return jsonify({'error': 'Email format not allowed\
+        an email must not start with @ or . character next to each other'}), 400
+    elif ".@" in user_email:
+        return jsonify({'error': 'Email format not allowed\
+        an email must not start with @ or . character next to each other'}), 400
+
+
+def validate_question(question_title):
+    check_duplicate_title = user_actions_object.get_user_name(question_title)
+
+    if len(check_duplicate_title) > 0:
+        return jsonify({'message': 'question already asked',
+        'goto': check_duplicate_title[2]}), 409
+    elif len(question_title) < 6:
+        return jsonify({'error': 'question title can not be less than six\
+            characters'}), 400
+    elif question_title.isdigit():
+        return jsonify({'error': 'question format not allowed\
+        an email can not only have numbers'}), 400
+
+
